# config/swagger.json
{
   "openapi":"3.0.0",
    "info":{
        "title":"API Documentation",
        "version":"1.0.0"
    },
    "components":{
        "securitySchemes":{
            "Bearer":{
                "type":"http",
                "scheme":"bearer",
                "bearerFormat":"JWT"
            }
        },
        "schemas":{
            "Event":{
                "type":"object",
                "properties":{
                    "eventname":{
                        "type":"string"
                    },
                    "date":{
                        "type":"string",
                        "format":"date"
                    },
                    "time":{
                        "type":"string"
                    },
                    "location":{
                        "type":"string"
                    },
                    "image":{
                        "type":"string"
                    }
                }
            },
            "Auth":{
                "type":"object",
                    "properties":{
                        "email":{
                            "type":"string",
                            "format":"email"
                        },
                        "password":{
                            "type":"string",
                            "format":"password"
                        }
             }
            }
        }
    },
    "paths":{
        "/event/list/{limit}/{offset}":{
            "get":{
                "tags":[
                    "Event"
                ],
                "summary":"Get list limit event",
                "security":[{ "Bearer": [] }],
                "parameters":[
                    {
                        "name":"limit",
                        "in":"path",
                        "description":"Limit list event",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    },
                    {
                        "name":"offset",
                        "in":"path",
                        "description":"Offset list user",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"List Event"
                    }
                }
            }
        },
        "/event":{
            "post":{
                "tags":[
                    "Event"
                ],
                "summary":"Create new event",
                "security":[{ "Bearer": [] }],
                    "requestBody":{
                        "required":true,
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref":"#/components/schemas/Event"
                                }
                            }
                        }
                    },
                "responses":{
                    "201":{
                        "description":"Event created"
                    }
                }
            }
        },
        "/event/{id}":{
            "get":{
                "tags":[
                    "Event"
                ],
                "summary":"Get event by id",
                "security":[{ "Bearer": [] }],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "description":"ID of event",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Event Found"
                    }
                }
            },
            "put":{
                "tags":[
                    "Event"
                ],
                "summary":"Update event",
                "security":[{ "Bearer": [] }],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "description":"ID of event",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/Event"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"Event Updated"
                    }
                }
            },
            "delete":{
                "tags":[
                    "Event"
                ],
                "summary":"Delete event",
                "security":[{ "Bearer": [] }],
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "description":"ID of event",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Event Deleted"
                    }
                }
            }
        },
        "/auth/signin":{
            "post":{
                "tags":[
                    "Auth"
                ],
                "summary":"Login user",
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/Auth"
                            }
                        }
                    }
            
        
                },
                "responses":{
                    "201":{
                        "description":"Token"
                    }
                }
            }
            },
        "/auth":{
            "get":{
                "tags":[
                    "Auth"
                ],
                "summary":"Get list user login",
                "responses":{
                    "200":{
                        "description":"List User"
                    }
                }
            },
            "post":{
                "tags":[
                    "Auth"
                ],
                "summary":"Create user login",
                "requestBody":{
                    "required":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "$ref":"#/components/schemas/Auth"
                            }
                        }
                    }
                },
                "responses":{
                    "201":{
                        "description":"User Created"
                    }
                }
            }
        },
        "/auth/{id}":{
            "get":{
                "tags":[
                    "Auth"
                ],
                "summary":"Get user login by id",
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "description":"ID of user",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User Found"
                    }
                }
            },
            "delete":{
                "tags":[
                    "Auth"
                ],
                "summary":"Delete user login",
                "parameters":[
                    {
                        "name":"id",
                        "in":"path",
                        "description":"ID of user",
                        "required":true,
                        "schema":{
                            "type":"string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User Deleted"
                    }
                }
            }
        }
    }
}